+++ Carbon.IIS/Functions/Get-CIisCollection.ps1
@@ -5,18 +5,18 @@ function Get-CIisCollection
     Gets an IIS configuration ollection.
 
     .DESCRIPTION
-    The `Get-CIisCollection` function gets the specified IIS collection. Pass the collection's IIS configuration section
-    path to the `SectionPath` parameter.
-
-    If the configuration section given by `SectionPath` is not a collection, pass the name of the collection to the
-    `Name` parameter.
-
-    If the collection needed is for a website, application, virtual directory, or directory, pass the path to that
-    location to the `LocatianPath` parameter.
+    The `Get-CIisCollection` function gets an IIS configuration element as a collection. Pass the collection's IIS
+    configuration section path to the `SectionPath` parameter. If the collection is actually a child element of the
+    configuration section, pass the name of the child element collection to the `Name` parameter. To get the section for
+    a specific site, directory, application, or virtual directory, pass its location path to the `LocationPath`
+    parameter.
 
     You can pass an instance of a `[Microsoft.Web.Administration.ConfigurationElement]` to the `ConfigurationElement`
     parameter to return that element as a collection, or, with the `Name` parameter, get a named collection under that
     configuration element.
+
+    This function returns a configuration element collection object. To get the items from the collection, use
+    `Get-CIisCollectionItem`.
 
     .EXAMPLE
     $collection = Get-CIisCollection -LocationPath 'SITE_NAME' -SectionPath 'system.webServer/httpProtocol/' -Name 'customHeaders'
@@ -26,18 +26,23 @@ function Get-CIisCollection
     #>
     [CmdletBinding()]
     param(
+        # The `[Microsoft.Web.Administration.ConfigurationElement]` object to get as a collection or the parent element
+        # of the collection element to get. If this is the parent element, pass the name of the child element collection
+        # to the `CollectionName` parameter.
         [Parameter(Mandatory, ParameterSetName='Direct')]
         [ConfigurationElement] $ConfigurationElement,
 
-        # The path for the configuration section that points to the collection
+        # The configuration section path of the collection, or, if the configuration section is a parent of the
+        # collection, the configuration section path to the parent configuration section. If the configuration section
+        # is the parent of the collection, pass the collection name to the `CollectionName` parameter.
         [Parameter(Mandatory, ParameterSetName='ByPath')]
         [String] $SectionPath,
 
-        # The name of the site where the collection belongs
+        # The location path to the site, directory, application, or virtual directory to configure.
         [Parameter(ParameterSetName='ByPath')]
         [String] $LocationPath,
 
-        # The name of the collection
+        # The collection's name.
         [String] $Name
     )