+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 `clear` element to the collection.
 * Function `Suspend-CIisAutoCommit` to stop Carbon.IIS from committing changes.
 * Function `Resume-CIisAutoCommit` to start Carbon.IIS commiting changes again.
+* Function `Get-CIisCollectionItem` for getting the items from an IIS configuration collection.
 
 ### Fixed